Critical trace updates (#6)
* critical trace updates

- better handling of OMNITRACE_USE_PERFETTO in omnitrace-critical-trace exe
- changed some data types in `critical_trace::entry`
- added device ids to critical trace entries
- added process ids to critical trace entries
- added packing to critical trace entries

* Update timemory submodule

+++ source/lib/omnitrace/library/runtime.hpp
@@ -73,8 +73,8 @@ get_cpu_cid();
 unique_ptr_t<std::vector<uint64_t>>&
 get_cpu_cid_stack(int64_t _tid = threading::get_id(), int64_t _parent = 0);
 
-using cpu_cid_data_t       = std::tuple<uint64_t, uint64_t, uint16_t>;
-using cpu_cid_pair_t       = std::tuple<uint64_t, uint16_t>;
+using cpu_cid_data_t       = std::tuple<uint64_t, uint64_t, uint32_t>;
+using cpu_cid_pair_t       = std::tuple<uint64_t, uint32_t>;
 using cpu_cid_parent_map_t = std::unordered_map<uint64_t, cpu_cid_pair_t>;
 
 unique_ptr_t<cpu_cid_parent_map_t>&
